+++ docs/history.rst
@@ -6,6 +6,7 @@ Version History
 - Add :func:`sprockets.mixins.mediatype.add_transcoder`
 - Add :class:`sprockets.mixins.mediatype.transcoders.JSONTranscoder`
 - Add :class:`sprockets.mixins.mediatype.transcoders.MsgPackTranscoder`
+- Content type parameters are now passed to transcoder ``from_bytes`` methods
 
 `1.0.4`_ (14 Sep 2015)
 ----------------------